Move `VertexId` to `subgraph` module
