+++ apps/hash-graph/lib/graph/src/shared/identifier.rs
@@ -16,7 +16,7 @@ use crate::{
     identifier::{
         knowledge::EntityId,
         ontology::OntologyTypeVersion,
-        time::{LeftClosedTemporalInterval, Timestamp, VariableAxis},
+        time::{LeftClosedTemporalInterval, VariableAxis},
     },
     subgraph::identifier::{EntityVertexId, OntologyTypeVertexId},
 };
@@ -44,55 +44,6 @@ impl From<OntologyTypeVertexId> for GraphElementVertexId {
 impl From<EntityVertexId> for GraphElementVertexId {
     fn from(id: EntityVertexId) -> Self {
         Self::KnowledgeGraph(id)
-    }
-}
-
-pub trait VertexId {
-    type BaseId;
-    type RevisionId;
-
-    fn new(base_id: Self::BaseId, revision_id: Self::RevisionId) -> Self;
-    fn base_id(&self) -> &Self::BaseId;
-    fn revision_id(&self) -> Self::RevisionId;
-}
-
-impl VertexId for OntologyTypeVertexId {
-    type BaseId = BaseUrl;
-    type RevisionId = OntologyTypeVersion;
-
-    fn new(base_id: Self::BaseId, revision_id: Self::RevisionId) -> Self {
-        Self {
-            base_id,
-            revision_id,
-        }
-    }
-
-    fn base_id(&self) -> &Self::BaseId {
-        &self.base_id
-    }
-
-    fn revision_id(&self) -> Self::RevisionId {
-        self.revision_id
-    }
-}
-
-impl VertexId for EntityVertexId {
-    type BaseId = EntityId;
-    type RevisionId = Timestamp<VariableAxis>;
-
-    fn new(base_id: Self::BaseId, revision_id: Self::RevisionId) -> Self {
-        Self {
-            base_id,
-            revision_id,
-        }
-    }
-
-    fn base_id(&self) -> &Self::BaseId {
-        &self.base_id
-    }
-
-    fn revision_id(&self) -> Self::RevisionId {
-        self.revision_id
     }
 }
