+++ apps/hash-graph/lib/graph/src/subgraph/identifier/vertex.rs
@@ -8,11 +8,32 @@ use crate::identifier::{
     time::{Timestamp, VariableAxis},
 };
 
+pub trait VertexId {
+    type BaseId;
+    type RevisionId;
+
+    fn base_id(&self) -> &Self::BaseId;
+    fn revision_id(&self) -> Self::RevisionId;
+}
+
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct OntologyTypeVertexId {
     pub base_id: BaseUrl,
     pub revision_id: OntologyTypeVersion,
+}
+
+impl VertexId for OntologyTypeVertexId {
+    type BaseId = BaseUrl;
+    type RevisionId = OntologyTypeVersion;
+
+    fn base_id(&self) -> &Self::BaseId {
+        &self.base_id
+    }
+
+    fn revision_id(&self) -> Self::RevisionId {
+        self.revision_id
+    }
 }
 
 impl From<VersionedUrl> for OntologyTypeVertexId {
@@ -29,4 +50,17 @@ impl From<VersionedUrl> for OntologyTypeVertexId {
 pub struct EntityVertexId {
     pub base_id: EntityId,
     pub revision_id: Timestamp<VariableAxis>,
+}
+
+impl VertexId for EntityVertexId {
+    type BaseId = EntityId;
+    type RevisionId = Timestamp<VariableAxis>;
+
+    fn base_id(&self) -> &Self::BaseId {
+        &self.base_id
+    }
+
+    fn revision_id(&self) -> Self::RevisionId {
+        self.revision_id
+    }
 }
